fix(loyalty): fix arithmetic bugs in point earn rate and discount overflow
Fix integer division causing zero earn rate, fix discount
equivalent formula, widen types to prevent overflow, and
fix downstream type mismatch in checkout service.

+++ Ecommerce/Application/DTOs/Loyalty/GetLoyaltyBalanceResponse.cs
@@ -2,6 +2,6 @@ namespace Application.DTOs.Loyalty
 {
     public sealed record GetLoyaltyBalanceResponse(
         int Balance,
-        int DiscountEquivalent,
+        long DiscountEquivalent,
         DateTime LastUpdated);
 }
